Create State-Sector Archive Generator
Created a custom Jekyll plugin that creates an individual archive page
for each state and all sectors. For example, for the state of “Assam”,
there is now an archive page for “Assam Energy” and “Assam Health”. The
generator runs on the build process, so any time new states or sectors
are created, they will be added.